Check for `.panel-collapse`, too.
This reverts back to the original selectors using `.parent()`.

The reason the first version of this ended up failing on the JS docs was
because it was looking for `.parent()` instead of filtering the `.parents()`.
In the JS docs' accordion example, `.panel-body` is wrapped with a
`.panel-collapse` div, so its direct parent wasn't `.panel`.

So I'm wondering if I go back to the original `.parent()` check instead of
the filtered `.parents()`. It'll be much more strict this way, though. But
it also catches the stray `<form>` mentioned in twbs/bootstrap#14300.